+++ tests/test_dynamic_graph_cpo_voem.cpp
@@ -1192,7 +1192,7 @@ TEST_CASE("voem CPO integration", "[dynamic_graph][voem][cpo][integration]") {
         visited[start] = true;
         order.push_back(start);
         
-        std::vector<uint32_t> queue = {start};
+        std::vector<uint32_t> queue = {static_cast<uint32_t>(start)};
         while (!queue.empty()) {
             auto uid = queue.front();
             queue.erase(queue.begin());
